chore(refactor): move stop execution and walk dom to utils
